+++ flink/src/main/java/org/apache/sedona/flink/Catalog.java
@@ -30,7 +30,9 @@ public class Catalog {
                 new Functions.ST_Distance(),
                 new Functions.ST_Transform(),
                 new Functions.ST_FlipCoordinates(),
-                new Functions.ST_GeoHash()
+                new Functions.ST_GeoHash(),
+                new Functions.ST_PointOnSurface(),
+                new Functions.ST_Reverse()
         };
     }