+++ flink/src/test/java/org/apache/sedona/flink/FunctionTest.java
@@ -64,4 +64,19 @@ public class FunctionTest extends TestBase{
         );
         assertEquals(first(pointTable).getField(0), Optional.of("s0000"));
     }
+
+    @Test
+    public void testPointOnSurface() {
+        Table pointTable = createPointTable_real(testDataSize);
+        Table surfaceTable = pointTable.select(call(Functions.ST_PointOnSurface.class.getSimpleName(), $(pointColNames[0])));
+        Geometry result = (Geometry) first(surfaceTable).getField(0);
+        assertEquals("POINT (32 -118)", result.toString());
+    }
+
+    public void testReverse() {
+        Table polygonTable = createPolygonTable(1);
+        Table ReversedTable = polygonTable.select(call(Functions.ST_Reverse.class.getSimpleName(), $(polygonColNames[0])));
+        Geometry result = (Geometry) first(ReversedTable).getField(0);
+        assertEquals("POLYGON ((-0.5 -0.5, 0.5 -0.5, 0.5 0.5, -0.5 0.5, -0.5 -0.5))", result.toString());
+    }
 }