+++ sql/src/main/scala/org/apache/spark/sql/sedona_sql/expressions/Functions.scala
@@ -1528,3 +1528,47 @@ case class ST_Multi(inputExpressions: Seq[Expression]) extends UnaryGeometryExpr
     Collect.createMultiGeometry(Seq(geometry)).toGenericArrayData
   }
 }
+
+/**
+ * Returns a POINT guaranteed to lie on the surface.
+ *
+ * @param inputExpressions Geometry
+ */
+case class ST_PointOnSurface(inputExpressions: Seq[Expression])
+  extends UnaryGeometryExpression with CodegenFallback {
+  assert(inputExpressions.length == 1)
+
+  override protected def nullSafeEval(geometry: Geometry): Any = {
+    new GenericArrayData(GeometrySerializer.serialize(GeomUtils.getInteriorPoint(geometry)))
+  }
+
+  override def dataType: DataType = GeometryUDT
+
+  override def children: Seq[Expression] = inputExpressions
+
+  protected def withNewChildrenInternal(newChildren: IndexedSeq[Expression]) = {
+    copy(inputExpressions = newChildren)
+  }
+}
+
+/**
+ * Returns the geometry with vertex order reversed
+ *
+ * @param inputExpressions
+ */
+case class ST_Reverse(inputExpressions: Seq[Expression])
+  extends UnaryGeometryExpression with CodegenFallback {
+  assert(inputExpressions.length == 1)
+
+  override protected def nullSafeEval(geometry: Geometry): Any = {
+    new GenericArrayData(GeometrySerializer.serialize(geometry.reverse()))
+  }
+
+  override def dataType: DataType = GeometryUDT
+
+  override def children: Seq[Expression] = inputExpressions
+
+  protected def withNewChildrenInternal(newChildren: IndexedSeq[Expression]) = {
+    copy(inputExpressions = newChildren)
+  }
+}